feat(journey-app): delete webauthn device from AM using device client
